+++ R/brokenstick.R
@@ -19,6 +19,7 @@
 #'
 #' @param knots Optional, but recommended. Numerical vector with the
 #' locations of the internal knots to be placed on the values of the predictor.
+#' The function sorts the internal knots in increasing order.
 #'
 #' @param boundary Optional, but recommended. Numerical vector of
 #' length 2 with the left and right boundary knot. The `boundary`
@@ -33,7 +34,10 @@
 #' specifying `k = 1` puts a knot at the 50th quantile (median),
 #' setting `k = 3` puts knots at the 25th, 50th and 75th quantiles,
 #' and so on. If the user specifies both `k` and `knots` arguments
-#' then `knots` takes precedence.
+#' then `knots` takes precedence. The default is `k = 5`, so if the user
+#' does not specify any of `knots`, `boundary` or `k`, then the knots
+#' will be at the 16th, 33th, 50th, 66th and 84th quantile of the
+#' predictor.
 #'
 #' @param degree the degree of the spline. The broken stick model
 #' requires linear splines, so the default is `degree = 1`.
@@ -119,7 +123,7 @@
 #' data <- smocc_200[1:1198, ]
 #'
 #' # using kr method, default
-#' f1 <- brokenstick(hgt_z ~ age | id, data, knots = 0:3, seed = 123)
+#' f1 <- brokenstick(hgt_z ~ age | id, data, knots = 0:2, seed = 123)
 #' plot(f1, data, n_plot = 9)
 #'
 #' # study sampling behaviour of the sigma2 parameter with coda
@@ -128,7 +132,7 @@
 #' acfplot(f1$mod$sigma2)
 #'
 #' # using lmer method
-#' f2 <- brokenstick(hgt_z ~ age | id, data, knots = 0:3, method = "lmer")
+#' f2 <- brokenstick(hgt_z ~ age | id, data, knots = 0:2, method = "lmer")
 #' plot(f2, data, n_plot = 9)
 #'
 #' # drill down into merMod object with standard diagnostics in lme4
@@ -151,7 +155,7 @@ brokenstick <- function(formula,
                         data,
                         knots = NULL,
                         boundary = NULL,
-                        k = NULL,
+                        k = 5L,
                         degree = 1L,
                         method = c("kr", "lmer"),
                         control = set_control(method = method, ...),