+++ inst/doc/mainfunctions.R
@@ -1,9 +1,10 @@
 ## ----setup, include=FALSE-----------------------------------------------------
+old <- options(digits = 3)
 knitr::opts_chunk$set(echo = TRUE, 
                       fig.width = 7, fig.height = 3,
                       dev = "png")
 
-## ----message = FALSE----------------------------------------------------------
+## ----message = FALSE, warning = FALSE-----------------------------------------
 require("brokenstick")
 require("dplyr")
 library("ggplot2")
@@ -34,27 +35,29 @@ ggplot(data[1:500, ], aes(x = age, y = hgt_z, group = id, color = as.factor(id))
 
 ## ----figure1, warning = FALSE-------------------------------------------------
 set.seed(123)
-fit <- brokenstick(hgt ~ age | id, data)
+fit <- brokenstick(hgt ~ age | id, data, knots = c(0, 2))
 ids <- c(10001, 10005, 10022)
-plot(fit, group = ids, what = "all",
+plot(fit, group = ids, 
      xlab = "Age (years)", ylab = "Length (cm)")
 
 ## ----zscore, warning = FALSE--------------------------------------------------
-fit0 <- brokenstick(hgt_z ~ age | id, data)
-plot(fit0, group = ids, what = "all", 
+fit0 <- brokenstick(hgt_z ~ age | id, data, knots = c(0, 2))
+plot(fit0, group = ids,
      xlab = "Age (years)", ylab = "Length (SDS)")
 
 ## ----plotfit2, cache = TRUE, warning=FALSE------------------------------------
-fit2 <- brokenstick(hgt_z ~ age | id, data = data, knots = 0:3)
+fit2 <- brokenstick(hgt_z ~ age | id, data = data, knots = c(0, 1, 2))
 plot(fit2, group = ids, xlab = "Age (years)", ylab = "Length (SDS)")
 
 ## -----------------------------------------------------------------------------
-fit2
+summary(fit2)
 
 ## ----fit9, cache = TRUE, warning = FALSE--------------------------------------
 knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24)/12, 4)
-fit9 <- brokenstick(hgt_z ~ age | id, data = data, 
-                    knots = knots, boundary = c(0, 3))
+fit9 <- brokenstick(hgt_z ~ age | id, data = data, knots = knots)
+
+## -----------------------------------------------------------------------------
+summary(fit9)
 
 ## ----plotfit9, echo = FALSE, warning = FALSE----------------------------------
 plot(fit9, group = ids, xlab = "Age (years)", ylab = "Length (SDS)")
@@ -65,12 +68,12 @@ head(p1)
 identical(nrow(data), nrow(p1))
 
 ## -----------------------------------------------------------------------------
-p2 <- predict(fit2, x = "knots")
+p2 <- predict(fit2, x = "knots", include_data = FALSE)
 head(p2)
 nrow(p2)
 
 ## -----------------------------------------------------------------------------
-p3 <- predict(fit2, x = "knots", strip_data = FALSE)
+p3 <- predict(fit2, x = "knots")
 table(p3$.source)
 
 ## -----------------------------------------------------------------------------
@@ -87,7 +90,7 @@ subj <- data %>%
 head(subj, 3)
 
 ## ----predictatknots-----------------------------------------------------------
-bs <- predict(fit9, x = "knots", shape = "wide")
+bs <- predict(fit9, x = "knots", shape = "wide", include_data = FALSE)
 data <- bind_cols(subj, select(bs, -id))
 head(data, 3)
 
@@ -98,4 +101,7 @@ summary(fit1_lm)
 ## ----lm2----------------------------------------------------------------------
 fit2_lm <- lm(`2` ~ sex + ga + I(bw / 1000) + `0`, data = data)
 summary(fit2_lm)
+
+## ----echo=FALSE---------------------------------------------------------------
+options(old)